arnekt_cogTour/requirements.txt
jsonpickle
tqdm
google-cloud
flask
speechRecognition
google-api-python-client
